pyproject.toml
[tool.poetry]
name = "rest_framework_securejwt"
version = "0.1.0"
description = "Authentication with access token and refresh token"
authors = ["Applichic <gbelouin@applichic.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
PyJWT = "^1.7"
djangorestframework = "^3.10"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
